🔨 Refactor: display none 을 사용하여 모바일에 따른 네비게이션바 리팩토링
